fix: tipo de dato de montoCreditoDebito consulta IATA

+++ sicf-consultas-generales/src/main/java/ob/unibanca/sicf/consultasgenerales/model/criterio/incomingIATA/CriterioBusquedaIncomingIATA.java
@@ -81,7 +81,7 @@ public class CriterioBusquedaIncomingIATA {
 	Integer ordenFechaEmisionBoleto;
 	Integer ordenPeriodoBSP;
 	String	ordenFacturacionADP;
-	Long 	ordenMontoCreditoDebito;
+	Double 	ordenMontoCreditoDebito;
 	Long ordenCodigoAutorizacion;
 	Integer ordenSequenceNumber;
 	Integer ordenInvoiceDate;
