+++ sicf-consultas-generales/src/main/java/ob/unibanca/sicf/consultasgenerales/model/incomingIATA/IncomingIATA.java
@@ -34,7 +34,7 @@ public class IncomingIATA {
 	String  numeroFactura;
 	String	nombrePasajero;
 	String codigoAutorizacion;
-	Long 	montoCreditoDebito;
+	Double 	montoCreditoDebito;
 	String  codigoMonedaIso;
